pyproject.toml
[project]
name = "mcp-rag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.12.0",
    "lxml>=4.9.0",
    "mcp[cli]>=1.12.0",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "pathlib2>=2.3.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "chromadb>=0.4.0",
    "sentence-transformers>=2.2.0",
    "markdown>=3.5.0",
]
